+++ packages/lumx-core/src/js/components/Checkbox/index.tsx
@@ -1,6 +1,14 @@
 import { mdiCheck, mdiMinus } from '@lumx/icons';
 
-import type { JSXElement, LumxClassName, HasTheme, HasAriaDisabled, HasClassName, CommonRef } from '../../types';
+import type {
+    JSXElement,
+    LumxClassName,
+    HasTheme,
+    HasAriaDisabled,
+    HasDisabled,
+    HasClassName,
+    CommonRef,
+} from '../../types';
 import { classNames } from '../../utils';
 import { Icon } from '../Icon';
 import { InputLabel } from '../InputLabel';
@@ -14,7 +22,7 @@ export const INTERMEDIATE_STATE = 'intermediate';
 /**
  * Defines the props of the component.
  */
-export interface CheckboxProps extends HasTheme, HasClassName, HasAriaDisabled {
+export interface CheckboxProps extends HasTheme, HasClassName, HasAriaDisabled, HasDisabled {
     /** Helper text. */
     helper?: string;
     /** Native input id property. */
@@ -23,8 +31,6 @@ export interface CheckboxProps extends HasTheme, HasClassName, HasAriaDisabled {
     isChecked?: boolean | 'intermediate';
     /** @alias isChecked */
     checked?: boolean;
-    /** Whether the component is disabled or not. */
-    isDisabled?: boolean;
     /** Label text. */
     label?: JSXElement;
     /** Native input name property. */