+++ packages/lumx-core/src/js/components/RadioButton/index.tsx
@@ -1,4 +1,12 @@
-import type { JSXElement, LumxClassName, HasTheme, HasAriaDisabled, HasClassName, CommonRef } from '../../types';
+import type {
+    JSXElement,
+    LumxClassName,
+    HasTheme,
+    HasAriaDisabled,
+    HasDisabled,
+    HasClassName,
+    CommonRef,
+} from '../../types';
 import { classNames } from '../../utils';
 import { InputLabel } from '../InputLabel';
 import { InputHelper } from '../InputHelper';
@@ -6,7 +14,7 @@ import { InputHelper } from '../InputHelper';
 /**
  * Defines the props of the component.
  */
-export interface RadioButtonProps extends HasTheme, HasClassName, HasAriaDisabled {
+export interface RadioButtonProps extends HasTheme, HasClassName, HasAriaDisabled, HasDisabled {
     /** Helper text. */
     helper?: string;
     /** Native input id property. */
@@ -15,8 +23,6 @@ export interface RadioButtonProps extends HasTheme, HasClassName, HasAriaDisable
     isChecked?: boolean;
     /** @alias isChecked */
     checked?: boolean;
-    /** Whether the component is disabled or not. */
-    isDisabled?: boolean;
     /** Label content. */
     label?: JSXElement;
     /** Native input name property. */